+++ spec/support/test_schemas.rb
@@ -10,17 +10,17 @@ module Mono
     column  name: 'Name',
             value: ->(row, *) { row[:name] }
 
-    columns name: ['Pet 1', 'Pet 2', 'Pet 3'],
+    column  name: ['Pet 1', 'Pet 2', 'Pet 3'],
             value: ->(row, *) { row[:pets] }
 
-    columns lambda { |table|
+    columns do |table|
       table[:questions].map do |question|
         {
           name: question[:id],
           value: ->(row, *) { row[:answers][question[:id]] }
         }
       end
-    }
+    end
   end
 
   module WithKeys
@@ -35,11 +35,11 @@ module Mono
               key: :name,
               value: ->(row, *) { row[:name] }
 
-      columns name: ['Pet 1', 'Pet 2', 'Pet 3'],
+      column  name: ['Pet 1', 'Pet 2', 'Pet 3'],
               key: %i[pet1 pet2 pet3],
               value: ->(row, *) { row[:pets] }
 
-      columns lambda { |table|
+      columns do |table|
         table[:questions].map do |question|
           {
             name: question[:id],
@@ -47,7 +47,7 @@ module Mono
             value: ->(row, *) { row[:answers][question[:id]] }
           }
         end
-      }
+      end
     end
   end
 end
@@ -66,21 +66,21 @@ module Micro
   class PetTableSchema
     include TableStructure::Schema
 
-    columns name: ['Pet 1', 'Pet 2', 'Pet 3'],
+    column  name: ['Pet 1', 'Pet 2', 'Pet 3'],
             value: ->(row, *) { row[:pets] }
   end
 
   class QuestionTableSchema
     include TableStructure::Schema
 
-    columns lambda { |table|
+    columns do |table|
       table[:questions].map do |question|
         {
           name: question[:id],
           value: ->(row, *) { row[:answers][question[:id]] }
         }
       end
-    }
+    end
   end
 
   module WithKeys
@@ -99,7 +99,7 @@ module Micro
     class PetTableSchema
       include TableStructure::Schema
 
-      columns name: ['Pet 1', 'Pet 2', 'Pet 3'],
+      column  name: ['Pet 1', 'Pet 2', 'Pet 3'],
               key: %i[pet1 pet2 pet3],
               value: ->(row, *) { row[:pets] }
     end